OT-2776 Switch from exit_pinmode to GTLoader.out_pinmode
