+++ keentools/geotracker/settings.py
@@ -56,7 +56,7 @@ def update_camobj(self, context: Any) -> None:
     if not self.camobj:
         settings = get_gt_settings()
         if settings.pinmode:
-            settings.force_out_pinmode = True
+            GTLoader.out_pinmode()
             return
     GTLoader.update_all_viewport_shaders()
 
@@ -67,7 +67,7 @@ def update_geomobj(self, context: Any) -> None:
     if not self.geomobj:
         settings = get_gt_settings()
         if settings.pinmode:
-            settings.force_out_pinmode = True
+            GTLoader.out_pinmode()
             return
     GTLoader.update_all_viewport_shaders()
 
@@ -234,9 +234,6 @@ class GTSceneSettings(bpy.types.PropertyGroup):
     pinmode: bpy.props.BoolProperty(name='Pinmode status', default=False)
     move_pin_mode: bpy.props.BoolProperty(name='Move pin mode status', default=False)
     pinmode_id: bpy.props.StringProperty(name='Unique pinmode ID')
-
-    force_out_pinmode: bpy.props.BoolProperty(name='Pin Mode Out', default=False)
-    license_error: bpy.props.BoolProperty(name='License Error', default=False)
 
     geotrackers: bpy.props.CollectionProperty(type=GeoTrackerItem, name='GeoTrackers')
     current_geotracker_num: bpy.props.IntProperty(name='Current Geotracker Number', default=-1)
